Lot of updates
- Compose plugin updated
- Removed preview-mock module
- Added pressure & temp unit selection

+++ core/src/main/java/com/masselis/tpmsadvanced/interfaces/viewmodel/SensorFavouriteViewModel.kt
@@ -28,7 +28,7 @@ class SensorFavouriteViewModel @AssistedInject constructor(
         object Empty : State()
 
         @Parcelize
-        data class RequestBond(val id: Short) : State()
+        data class RequestBond(val id: Int) : State()
     }
 
     private val mutableStateFlow = savedStateHandle
